refactor(corporativo): decompose CorpContactSection into focused subcomponents
Split the 370-line giant component into three focused files to resolve
the react-doctor no-giant-component warning:

- CorpContactSection (orchestrator, ~24 lines): computes whatsappUrl once
- CorpContactInfo (presentational): left panel with contact details and social links
- CorpContactForm (stateful): form state, validation, dual submit (SF + n8n)

All conversion-critical behavior preserved verbatim: dual Promise.all submit,
isLocallySubmitting ref guard, data-tracking/data-contact-intent attributes.
